+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project follows [these versioning rules](https://sagegerard.com/edition-rev
 ## Unreleased
 
 - Rename to Denxi
+- Allow per-artifact installations
 - Change default branch from `master` to `default`
 - Allow use of custom cryptographic backends to unblock users when the
   bundled library doesn't load.